:truck: move channel to channels folder
